bigbrother/speye/requirements.txt
wheel
requests
pandas
sqlalchemy
psycopg2-binary
python-dotenv
pytz